expand goto_programt and goto_functionst templates
The goto_program_template and the goto_functions_template each only have one
instance, but make it much harder to read the code.
