+++ src/cbmc/symex_coverage.cpp
@@ -23,6 +23,8 @@ Date: March 2016
 #include <util/cprover_prefix.h>
 #include <util/prefix.h>
 
+#include <langapi/language_util.h>
+
 #include <goto-programs/goto_functions.h>
 #include <goto-programs/remove_returns.h>
