+++ src/goto-programs/goto_inline_class.cpp
@@ -25,6 +25,8 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <util/std_code.h>
 #include <util/std_expr.h>
 
+#include <langapi/language_util.h>
+
 #include "remove_skip.h"
 #include "goto_inline.h"
